added start and stop button, modified cell object, consolidated helpers, added generation display
